+++ src/main.js
@@ -1,12 +1,12 @@
-import RouteInfo from "./view/site-info-route";
-import RouteCost from "./view/site-price";
-import SiteMenuView from "./view/site-menu";
-import SiteFilterView from "./view/site-filter";
-import SiteSortView from "./view/site-sort";
-import SiteDaysView from "./view/site-days";
-import SiteDayView from "./view/site-day";
-import SiteWayPointView from "./view/site-waypoint";
-import WayPointEdit from "./view/site-form-with-change";
+import RouteInfoComponent from "./view/route-info";
+import RouteCostComponent from "./view/route-price";
+import MenuViewComponent from "./view/menu";
+import FilterComponent from "./view/filter";
+import SortComponent from "./view/sort";
+import DaysComponent from "./view/days";
+import DayComponent from "./view/day";
+import WayPointComponent from "./view/waypoint";
+import WayPointEditComponent from "./view/waypoint-edit";
 import {generateWaypoint} from "./mock/waypoint";
 import {getDatesDuration, render} from "./mock/utils";
 
@@ -14,15 +14,15 @@ const WAYPOINT_COUNT = 15;
 
 // дополнительно
 const siteTripInMainElement = document.querySelector(`.trip-main`);
-render(siteTripInMainElement, new RouteInfo().getElement(), `afterbegin`);
+render(siteTripInMainElement, new RouteInfoComponent().getElement(), `afterbegin`);
 
 const siteTripInfoMainElement = siteTripInMainElement.querySelector(`.trip-info__main`);
-render(siteTripInfoMainElement, new RouteCost().getElement(), `afterend`);
+render(siteTripInfoMainElement, new RouteCostComponent().getElement(), `afterend`);
 //
 const tripControls = siteTripInMainElement.querySelector(`.trip-controls`);
 const tripMenu = tripControls.querySelector(`h2`);
-render(tripMenu, new SiteMenuView().getElement(), `afterend`);
-render(tripControls, new SiteFilterView().getElement());
+render(tripMenu, new MenuViewComponent().getElement(), `afterend`);
+render(tripControls, new FilterComponent().getElement());
 
 const renderRoutePoint = (routePointList, routePoint) => {
   const onRollupButtonClick = () => {
@@ -39,11 +39,11 @@ const renderRoutePoint = (routePointList, routePoint) => {
     routePointList.replaceChild(routePointComponent.getElement(), routePointEditComponent.getElement());
   };
 
-  const routePointComponent = new SiteWayPointView(routePoint);
+  const routePointComponent = new WayPointComponent(routePoint);
   const rollupButton = routePointComponent.getElement().querySelector(`.event__rollup-btn`);
   rollupButton.addEventListener(`click`, onRollupButtonClick);
 
-  const routePointEditComponent = new WayPointEdit(routePoint, false);
+  const routePointEditComponent = new WayPointEditComponent(routePoint, false);
   const editForm = routePointEditComponent.getElement().querySelector(`form`);
   editForm.addEventListener(`submit`, onEditFormSubmit);
   editForm.addEventListener(`reset`, onEditFormClose);
@@ -53,20 +53,20 @@ const renderRoutePoint = (routePointList, routePoint) => {
 
 const renderRouteTable = (tripEvents, routePoints) => {
   const tripSorting = tripEvents.querySelector(`h2`);
-  render(tripSorting, new SiteSortView().getElement(), `afterend`);
+  render(tripSorting, new SortComponent().getElement(), `afterend`);
 
-  const daysComponent = new SiteDaysView();
+  const daysComponent = new DaysComponent();
   render(tripEvents, daysComponent.getElement());
 
   let dayNumber = 0;
-  let dayComponent = new SiteDayView(dayNumber + 1, routePoints[0]);
+  let dayComponent = new DayComponent(dayNumber + 1, routePoints[0]);
   render(daysComponent.getElement(), dayComponent.getElement());
 
   routePoints.forEach((routePoint, index) => {
     if (index > 0) {
       if (getDatesDuration(routePoints[0].timeBegin, routePoint.timeBegin).daysBetween > dayNumber) {
         dayNumber = getDatesDuration(routePoints[0].timeBegin, routePoint.timeBegin).daysBetween;
-        dayComponent = new SiteDayView(dayNumber + 1, routePoint);
+        dayComponent = new DayComponent(dayNumber + 1, routePoint);
         render(daysComponent.getElement(), dayComponent.getElement());
       }
       renderRoutePoint(dayComponent.getElement().querySelector(`.trip-events__list`), routePoint);
