+++ src/view/days.js
@@ -1,21 +1,19 @@
 import {createElement} from "../mock/utils";
 
-const createSiteMenuTemplate = () => {
+const createDaysTemplate = () => {
   return (
-    `<nav class="trip-controls__trip-tabs  trip-tabs">
-    <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-    <a class="trip-tabs__btn" href="#">Stats</a>
-    </nav>`
+    `<ul class="trip-days">
+    </ul>`
   );
 };
 
-export default class SiteMenu {
+export default class Days {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createSiteMenuTemplate();
+    return createDaysTemplate();
   }
 
   getElement() {
