+++ src/view/menu.js
@@ -1,20 +1,21 @@
 import {createElement} from "../mock/utils";
 
-const createSitePriceTemplate = () => {
+const createSiteMenuTemplate = () => {
   return (
-    `<p class="trip-info__cost">
-              Total: €&nbsp;<span class="trip-info__cost-value">1230</span>
-            </p>`
+    `<nav class="trip-controls__trip-tabs  trip-tabs">
+    <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+    <a class="trip-tabs__btn" href="#">Stats</a>
+    </nav>`
   );
 };
 
-export default class RouteCost {
+export default class Menu {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createSitePriceTemplate();
+    return createSiteMenuTemplate();
   }
 
   getElement() {
